# [perf] Add simple benchmarks for size/perf overhead