--- a/bench/src/add/addon.c
+++ b/bench/src/add/addon.c
@@ -1,0 +1,47 @@
+/*
+ * Copied from
+ * https://github.com/nodejs/node-addon-examples/blob/main/src/1-getting-started/2_function_arguments/napi/addon.c
+ *
+ * The MIT License (MIT)
+ * Copyright (c) 2017 Node.js node-addon-examples collaborators
+ */
+
+#include <assert.h>
+#include <node_api.h>
+#include <stdio.h>
+
+static napi_value Add(napi_env env, napi_callback_info info) {
+  napi_status status;
+
+  size_t argc = 2;
+  napi_value args[2];
+  status = napi_get_cb_info(env, info, &argc, args, NULL, NULL);
+  assert(status == napi_ok);
+
+  double value0;
+  status = napi_get_value_double(env, args[0], &value0);
+  assert(status == napi_ok);
+
+  double value1;
+  status = napi_get_value_double(env, args[1], &value1);
+  assert(status == napi_ok);
+
+  napi_value sum;
+  status = napi_create_double(env, value0 + value1, &sum);
+  assert(status == napi_ok);
+
+  return sum;
+}
+
+#define DECLARE_NAPI_METHOD(name, func)                                        \
+  {name, 0, func, 0, 0, 0, napi_default, 0}
+
+napi_value Init(napi_env env, napi_value exports) {
+  napi_status status;
+  napi_property_descriptor addDescriptor = DECLARE_NAPI_METHOD("add", Add);
+  status = napi_define_properties(env, exports, 1, &addDescriptor);
+  assert(status == napi_ok);
+  return exports;
+}
+
+NAPI_MODULE(NODE_GYP_MODULE_NAME, Init)