--- a/bench/src/add/addon.zig
+++ b/bench/src/add/addon.zig
@@ -1,0 +1,10 @@
+const std = @import("std");
+const t = @import("tokota");
+
+comptime {
+    t.exportModule(@This());
+}
+
+pub fn add(a: f64, b: f64) f64 {
+    return a + b;
+}