--- a/bench/src/hello/addon.c
+++ b/bench/src/hello/addon.c
@@ -1,0 +1,31 @@
+/*
+ * Copied from
+ * https://github.com/nodejs/node-addon-examples/blob/main/src/1-getting-started/1_hello_world/napi/hello.c
+ *
+ * The MIT License (MIT)
+ * Copyright (c) 2017 Node.js node-addon-examples collaborators
+ */
+
+#include <assert.h>
+#include <node_api.h>
+
+static napi_value Method(napi_env env, napi_callback_info info) {
+  napi_status status;
+  napi_value world;
+  status = napi_create_string_utf8(env, "world", 5, &world);
+  assert(status == napi_ok);
+  return world;
+}
+
+#define DECLARE_NAPI_METHOD(name, func)                                        \
+  {name, 0, func, 0, 0, 0, napi_default, 0}
+
+static napi_value Init(napi_env env, napi_value exports) {
+  napi_status status;
+  napi_property_descriptor desc = DECLARE_NAPI_METHOD("hello", Method);
+  status = napi_define_properties(env, exports, 1, &desc);
+  assert(status == napi_ok);
+  return exports;
+}
+
+NAPI_MODULE(NODE_GYP_MODULE_NAME, Init)