--- a/bench/src/hello/addon.zig
+++ b/bench/src/hello/addon.zig
@@ -1,0 +1,9 @@
+const t = @import("tokota");
+
+comptime {
+    t.exportModule(@This());
+}
+
+pub fn hello() []const u8 {
+    return "world";
+}