Want to keep track of/try to trim some of the overhead introduced by the generic wrappers, inferred conversion, etc.

Comparing plain C addons to addons using the Tokota comptime framework and the raw `tokota.napi` bindings - also throwing in `node-addon-api` addons (cpp) for a reference high-level framework to compare against.

`node-addon-api` binary sizes seem off - need to look into why/how to get that trimmed down.

### MacOS
```console
➜ cd bench && node src/add/main.js
clk: ~3.18 GHz
cpu: Apple M1 Ultra
runtime: node 24.2.0 (arm64-darwin)

benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
c                             22.42 ns/iter  22.24 ns █
                      (22.15 ns … 42.93 ns)  25.71 ns █
                    (  4.16  b …  81.02  b)  16.19  b █▆▁▁▃▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁

cpp (node-addon-api)          31.91 ns/iter  31.73 ns  █
                      (31.33 ns … 58.76 ns)  35.43 ns  █
                    (  1.22  b …  88.11  b)  16.22  b ▂██▁▂▂▃▂▂▁▁▁▁▁▁▁▁▁▁▁▁

zig (tokota - ffi)            22.37 ns/iter  22.28 ns █▅
                      (22.17 ns … 45.76 ns)  25.00 ns ██
                    (  1.22  b …  88.11  b)  16.17  b ██▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

zig (tokota)                  24.65 ns/iter  24.41 ns █
                      (24.26 ns … 51.90 ns)  27.74 ns █
                    (  9.22  b …  80.11  b)  16.17  b ██▁▁▂▂▄▂▂▁▁▁▁▁▁▁▁▁▁▁▁

summary
  zig (tokota - ffi)
   1x faster than c
   1.1x faster than zig (tokota)
   1.43x faster than cpp (node-addon-api)

Binary sizes
------------
┌──────────────────────┬────────────┐
│ (index)              │ Size (KB)  │
├──────────────────────┼────────────┤
│ c                    │ 49.15625   │
│ zig (tokota ffi)     │ 49.1953125 │
│ zig (tokota)         │ 50.09375   │
│ cpp (node-addon-api) │ 278.296875 │
└──────────────────────┴────────────┘
```

### Linux (VM)
```console
❯ node src/add/[main.js](http://main.js/) 
clk: ~3.07 GHz
cpu: unknown
runtime: node 24.2.0 (arm64-linux)

benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
c                             59.83 ns/iter  60.98 ns  █                   
                     (57.21 ns … 137.62 ns)  76.19 ns ▄█                   
                    (  4.38  b … 108.87  b)  16.26  b ██▆▄▅▅▃▂▂▂▂▁▁▁▁▁▁▁▁▁▁

cpp (node-addon-api)          64.85 ns/iter  65.63 ns  █▂                  
                     (62.50 ns … 126.80 ns)  74.30 ns  ██▆                 
                    ( 16.09  b … 112.15  b)  16.31  b ████▆▇█▅▄▂▂▃▂▂▁▁▁▁▁▁▁

zig (tokota - ffi)            58.80 ns/iter  58.56 ns  █                   
                      (57.67 ns … 98.31 ns)  65.89 ns  █                   
                    ( 16.09  b … 112.15  b)  16.24  b ▄█▇▃▃▃▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁

zig (tokota)                  61.82 ns/iter  61.57 ns ▂█                   
                     (60.93 ns … 108.78 ns)  68.33 ns ██                   
                    (  1.31  b … 112.15  b)  16.24  b ██▄▃▃▄▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁

summary
  zig (tokota - ffi)
   1.02x faster than c
   1.05x faster than zig (tokota)
   1.1x faster than cpp (node-addon-api)

Binary sizes
------------
┌──────────────────────┬────────────┐
│ (index)              │ Size (KB)  │
├──────────────────────┼────────────┤
│ zig (tokota ffi)     │ 3.3125     │
│ c                    │ 4          │
│ zig (tokota)         │ 11.7578125 │
│ cpp (node-addon-api) │ 244.015625 │
└──────────────────────┴────────────┘
```